+++ tests/e2e/app-files/hello.R
@@ -1,5 +1,5 @@
 box::use(
-  shiny[
+  shiny[           # nolint
     actionButton,
     bootstrapPage,
     isolate,